backend/requirements.txt
fastapi>=0.132.0
uvicorn[standard]>=0.41.0
pydantic>=2.12.5
pydantic-settings>=2.13.1
python-dotenv>=1.2.1
openai>=2.23.0
anthropic>=0.83.0
httpx>=0.28.1
sse-starlette>=3.2.0
python-multipart>=0.0.22
pypdf>=6.7.2
python-docx>=1.2.0
openpyxl>=3.1.5
chardet>=6.0.0.post1
google-cloud-secret-manager>=2.22.0
